Simplify: move data/cache prep to aux module; remove distortion.
Remove hardcoded paths, global var; fix log error.
